+++ lib/src/videos/video.dart
@@ -9,9 +9,11 @@ import 'video_id.dart';
 
 part 'video.freezed.dart';
 
+typedef MusicData = ({String? song, String? artist, String? album, Uri? image});
+
 /// YouTube video metadata.
 @freezed
-class Video with _$Video {
+abstract class Video with _$Video {
   /// Video URL.
   String get url => 'https://www.youtube.com/watch?v=$id';
 
@@ -58,6 +60,10 @@ class Video with _$Video {
     /// Returns true if this is a live stream.
     //ignore: avoid_positional_boolean_parameters
     bool isLive, [
+    /// Music data such as song, artist, album, and image.
+    /// Empty if no data is available.
+    List<MusicData> musicData = const [],
+
     /// Used internally.
     /// Shouldn't be used in the code.
     @internal WatchPage? watchPage,
@@ -89,6 +95,7 @@ class Video with _$Video {
       UnmodifiableListView(keywords ?? const Iterable.empty()),
       engagement,
       isLive,
+      musicData,
       watchPage,
     );
   }
@@ -133,7 +140,11 @@ class Video with _$Video {
 
     /// Returns true if this is a live stream.
     //ignore: avoid_positional_boolean_parameters
-    bool isLive, [
+    bool isLive,
+
+    /// Music data such as song, artist, album, and image.
+    /// Empty if no data is available.
+    List<MusicData> musicData, [
     /// Used internally.
     /// Shouldn't be used in the code.
     @internal WatchPage? watchPage,